first commit
CareU (Hospital Management System, AIF) - Developed a Hospital Management application where admin can register patients and create appointments for them with the doctors as per the availability. Patients will have log of all their visits and prescriptions. The Visits and prescriptions of the patients will be maintained automatically in a log.